Consolidated Widget UI and Design
- Removed unnecessary files, functions, and inheritance
- Created desired version of widget UI and design
- Moved necessary functionality into a single class

Functionality was spread out to follow how Unreal solves similar problems however this was premature. Some desirable functionality  is not easily included in how Unreal had solved such problems which then demands a reassessment of the problem space and the current code structure. So for ease of change all functionality for the widget has been consolidated for the time being.