Initial Fuzzy Search Design
+ Designed and implemented initial fuzzy search
- Removed functions for previous blueprint implementation

The current approach is a simple bucket sort where we sort paths into buckets based on the individual paths EditDistance, i.e. an integer representing how many characters must be added, removed, or swapped for two strings to be identical. We then sort individual buckets using selection sort and the EditDistance again.
The design is potentially inefficient as EditDistance uses FString::Compare which must rely on a string comparison algorithm. A more optimal algorithm is already designed however this simpler and easier approach must be successfully implemented and tested first.